Change default of cache input to false
We've found that the docker build action's caching doesn't effectively
work when being called by workflows running on different workers (with
seperate architectures). Ultimately only one architecture gets cached,
and the other architecture's builds are uncached. Since the overall CI
speed is limited by the slowest job, it's actually faster not to cache
because we don't spend time uploading layers to a cache that doesn't get
used. Therefore it makes sense to change the default for the cache input
to `false`.

I suppose this might be considered a major change? although in practice
I don't think it has adverse effects for any user.